Fix : move HomePage Widget to home folder
